# Conflicts:
#	app/src/main/java/com/geridea/trentastico/Config.java
#	app/src/main/java/com/geridea/trentastico/gui/views/CourseTimesCalendar.java
#	app/src/main/java/com/geridea/trentastico/services/LessonsUpdaterService.java
#	app/src/main/java/com/geridea/trentastico/utils/time/CalendarUtils.java

+++ app/src/main/java/com/geridea/trentastico/gui/fragments/CalendarFragment.java
@@ -177,11 +177,11 @@ public class CalendarFragment extends IFragmentWithMenuItems {
             setView(view);
         }
 
-        private ArrayList<Integer> calculateLessonTypesToHide() {
-            ArrayList<Integer> lessonTypesToHideIds = new ArrayList<>();
+        private ArrayList<Long> calculateLessonTypesToHide() {
+            ArrayList<Long> lessonTypesToHideIds = new ArrayList<>();
             for (LessonType lessonType : lessonTypes) {
                 if (!lessonType.isVisible()) {
-                    lessonTypesToHideIds.add(lessonType.getId());
+                    lessonTypesToHideIds.add((long) lessonType.getId());
                 }
             }
